Inline recipe and create separate tests
